Updated_README
This is what I was able to do so far. Since we do not really have a problem defined yet I just went off of what we had talked about in the last meeting. I added a few more headings, however, they are completely optional. If you decide they are not necessary, feel free to delete them.